task 7 PR2: compiled all db queries into one query. Also refactored files with names file_controller to file.controller etc. Also extracted errors into own error controller.
