+++ app.js
@@ -1,10 +1,11 @@
 const express = require("express");
 const app = express();
 app.use(express.json());
+const {handleRootPathErrors, handleCustomErrors,handlePSQLErrors} = require('./controllers/errors.controller.js')
 
-const { getTopics } = require("./controllers/topics_controller.js");
-const {getArticleById, patchArticleById} = require("./controllers/articles_controller.js");
-const {getAllUsers} = require("./controllers/users_controller.js")
+const { getTopics } = require("./controllers/topics.controller.js");
+const {getArticleById, patchArticleById} = require("./controllers/articles.controller.js");
+const {getAllUsers} = require("./controllers/users.controller.js")
 
 
 
@@ -14,30 +15,10 @@ app.patch("/api/articles/:article_id", patchArticleById);
 app.get("/api/users",getAllUsers)
 
 
-app.get("/api/*", (req, res) => {
-  res.status(404).send({ msg: "Not Found" });
-});
+app.get("/api/*", handleRootPathErrors)
 
-app.use((err, req, res, next) => {
-  if (err.status) {
-    res.status(err.status).send({ msg: err.msg });
-  } else {
-    next(err);
-  }
-});
+app.use(handleCustomErrors)
 
-app.use((err, req, res, next) => {
-  if (err.code === "22P02") {
-    res.status(400).send({ msg: "Bad Request" });
-  } else {
-    next(err);
-  }
-});
-
-app.use((err, req, res, next) => {
-  if (err.code === "23502") {
-    res.status(422).send({ msg: "Incorrect Input" });
-  }
-});
+app.use(handlePSQLErrors)
 
 module.exports = app;
