+++ db/seeds/run-seed.js
@@ -1,6 +1,6 @@
 const devData = require('../data/development-data/index.js');
 const seed = require('./seed.js');
-const db = require('../connection.js');
+const db = require('../index.js');
 
 const runSeed = () => {
   return seed(devData).then(() => db.end());
